requirements.txt
google-generativeai
openai
rich
requests
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
dnspython
python-whois
reportlab
markdown2
gunicorn
uvicorn[standard]
